+++ 01-Basic/31-60/45-FirstPackage/A1/program.py
@@ -1,4 +1,4 @@
-from Pack01.program import FunctionIn01Pack
+from B1.program import FunctionIn01Pack
 
 FunctionIn01Pack()
 # Output:
